backend/requirements.txt
fastapi==0.136.1
uvicorn==0.46.0
google-generativeai==0.8.6
python-dotenv==1.2.2
